🎨 (context-module): Improve typing
