+++ packages/signer/context-module/src/ContextModule.ts
@@ -4,8 +4,8 @@ import {
   type ClearSignContextType,
 } from "@/shared/model/ClearSignContext";
 
-import { type TypedDataClearSignContext } from "./shared/model/TypedDataClearSignContext";
-import { type TypedDataContext } from "./shared/model/TypedDataContext";
+import { type TypedDataClearSignContext } from "./modules/ethereum/model/TypedDataClearSignContext";
+import { type TypedDataContext } from "./modules/ethereum/model/TypedDataContext";
 
 export interface ContextModule {
   getContexts<TInput>(
