+++ packages/signer/context-module/src/index.ts
@@ -29,6 +29,9 @@ export * from "./modules/ethereum/gated-signing/data/GatedDescriptorDataSource";
 export * from "./modules/ethereum/gated-signing/data/HttpGatedDescriptorDataSource";
 export * from "./modules/ethereum/gated-signing/domain/GatedSigningContextLoader";
 export * from "./modules/ethereum/gated-signing/domain/GatedSigningTypedDataContextLoader";
+export * from "./modules/ethereum/model/EthereumClearSignContext";
+export * from "./modules/ethereum/model/TypedDataClearSignContext";
+export * from "./modules/ethereum/model/TypedDataContext";
 export * from "./modules/ethereum/nft/data/HttpNftDataSource";
 export * from "./modules/ethereum/nft/data/NftDataSource";
 export * from "./modules/ethereum/nft/domain/NftContextFieldLoader";
@@ -69,6 +72,8 @@ export * from "./modules/multichain/reporter/domain/BlindSigningReporter";
 export * from "./modules/multichain/reporter/domain/DefaultBlindSigningReporter";
 export * from "./modules/multichain/reporter/model/BlindSigningEvent";
 export * from "./modules/multichain/reporter/model/BlindSigningModelId";
+export * from "./modules/solana/model/SolanaClearSignContext";
+export * from "./modules/solana/model/SolanaContextTypes";
 export * from "./modules/solana/owner-info/data/HttpOwnerInfoDataSource";
 export * from "./modules/solana/owner-info/data/OwnerInfoDataSource";
 export * from "./modules/solana/owner-info/domain/OwnerInfoContextLoader";
@@ -78,7 +83,4 @@ export * from "./shared/domain/ContextLoader";
 export * from "./shared/domain/ContextModuleChainID";
 export * from "./shared/model/ClearSignContext";
 export * from "./shared/model/GenericPath";
-export * from "./shared/model/SolanaContextTypes";
 export * from "./shared/model/TransactionSubset";
-export * from "./shared/model/TypedDataClearSignContext";
-export * from "./shared/model/TypedDataContext";
