+++ packages/signer/context-module/src/modules/ethereum/model/TypedDataClearSignContext.ts
