+++ packages/signer/context-module/src/modules/ethereum/typed-data/domain/TypedDataContextLoader.ts
@@ -1,5 +1,5 @@
-import type { TypedDataClearSignContext } from "@/shared/model/TypedDataClearSignContext";
-import type { TypedDataContext } from "@/shared/model/TypedDataContext";
+import type { TypedDataClearSignContext } from "@/modules/ethereum/model/TypedDataClearSignContext";
+import type { TypedDataContext } from "@/modules/ethereum/model/TypedDataContext";
 
 export interface TypedDataContextLoader {
   load(typedData: TypedDataContext): Promise<TypedDataClearSignContext>;
