+++ packages/signer/context-module/src/modules/solana/owner-info/data/OwnerInfoDataSource.ts
@@ -1,6 +1,6 @@
 import { type Either } from "purify-ts";
 
-import type { SolanaTransactionContext } from "@/shared/model/SolanaTransactionContext";
+import type { SolanaTransactionContext } from "@/modules/solana/model/SolanaTransactionContext";
 
 export type HttpOwnerInfoDataSourceResult = {
   tlvDescriptor: Uint8Array;
